feat: add repository list and detail view controllers
Resolves: none.
