+++ GitIt/View/Class/HistoryCollectionViewCell.swift
