+++ GitIt/View/Class/HistoryTableViewCell.swift
