+++ GitIt/View/Extension/UIColor.swift
@@ -1,0 +1,62 @@
+//
+//  UIColor.swift
+//  GitIt
+//
+//  Created by Loay Ashraf on 05/12/2021.
+//
+
+import Foundation
+import UIKit
+
+extension UIColor {
+
+    convenience init(rgb: UInt) {
+        self.init(rgb: rgb, alpha: 1.0)
+    }
+
+    convenience init(rgb: UInt, alpha: CGFloat) {
+        self.init(
+            red: CGFloat((rgb & 0xFF0000) >> 16) / 255.0,
+            green: CGFloat((rgb & 0x00FF00) >> 8) / 255.0,
+            blue: CGFloat(rgb & 0x0000FF) / 255.0,
+            alpha: CGFloat(alpha)
+        )
+    }
+    
+    convenience init?(hex: String) {
+        let r, g, b, a: CGFloat
+
+        if hex.hasPrefix("#") {
+            let start = hex.index(hex.startIndex, offsetBy: 1)
+            let hexColor = String(hex[start...])
+
+            if hexColor.count == 6 {
+                let scanner = Scanner(string: hexColor)
+                var hexNumber: UInt64 = 0
+                if scanner.scanHexInt64(&hexNumber) {
+                    r = CGFloat((hexNumber & 0xff0000) >> 16) / 255.0
+                    g = CGFloat((hexNumber & 0x00ff00) >> 8) / 255.0
+                    b = CGFloat((hexNumber & 0x0000ff)) / 255.0
+                    a = 1.0
+
+                    self.init(red: r, green: g, blue: b, alpha: a)
+                    return
+                }
+            } else if hexColor.count == 8 {
+                let scanner = Scanner(string: hexColor)
+                var hexNumber: UInt64 = 0
+                if scanner.scanHexInt64(&hexNumber) {
+                    r = CGFloat((hexNumber & 0xff000000) >> 24) / 255
+                    g = CGFloat((hexNumber & 0x00ff0000) >> 16) / 255
+                    b = CGFloat((hexNumber & 0x0000ff00) >> 8) / 255
+                    a = CGFloat(hexNumber & 0x000000ff) / 255
+
+                    self.init(red: r, green: g, blue: b, alpha: a)
+                    return
+                }
+            }
+        }
+        return nil
+    }
+    
+}
