+++ GitIt/View/Extension/UIView.swift
@@ -1,0 +1,28 @@
+//
+//  UIView.swift
+//  GitIt
+//
+//  Created by Loay Ashraf on 05/12/2021.
+//
+
+import Foundation
+import UIKit
+
+extension UIView {
+    
+    var cornerRadius: CGFloat {
+        get { return layer.cornerRadius }
+        set { layer.cornerRadius = newValue }
+    }
+    
+    var cornerCurve: CALayerCornerCurve {
+        get { return layer.cornerCurve }
+        set { layer.cornerCurve = newValue }
+    }
+    
+    var masksToBounds: Bool {
+        get { return layer.masksToBounds }
+        set { layer.masksToBounds = newValue }
+    }
+    
+}
