+++ GitIt/View/Interface Builder/Storyboard/Base.lproj/LaunchScreen.storyboard
