+++ GitIt/View/Interface Builder/XIB/HistoryCollectionViewCell.xib
@@ -1,0 +1,64 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="18122" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="18093"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="System colors in document resources" minToolsVersion="11.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" reuseIdentifier="HistoryCollectionViewCell" id="gTV-IL-0wX" customClass="HistoryCollectionViewCell" customModule="GitIt" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="74" height="100"/>
+            <autoresizingMask key="autoresizingMask"/>
+            <view key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center">
+                <rect key="frame" x="0.0" y="0.0" width="74" height="100"/>
+                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                <subviews>
+                    <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="jr7-cp-sPd" customClass="AsyncUIImageView" customModule="GitIt" customModuleProvider="target">
+                        <rect key="frame" x="5" y="5" width="64" height="64"/>
+                        <constraints>
+                            <constraint firstAttribute="height" constant="64" id="68A-I9-BZi"/>
+                            <constraint firstAttribute="width" constant="64" id="ASp-MZ-zCq"/>
+                        </constraints>
+                    </imageView>
+                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="center" lineBreakMode="tailTruncation" numberOfLines="2" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="QfU-Gs-seW">
+                        <rect key="frame" x="5" y="74" width="64" height="20"/>
+                        <constraints>
+                            <constraint firstAttribute="width" constant="64" id="XY5-ZH-suQ"/>
+                            <constraint firstAttribute="height" constant="20" id="g5N-52-e1Q"/>
+                        </constraints>
+                        <fontDescription key="fontDescription" type="system" pointSize="15"/>
+                        <nil key="textColor"/>
+                        <nil key="highlightedColor"/>
+                    </label>
+                </subviews>
+            </view>
+            <viewLayoutGuide key="safeArea" id="ZTg-uK-7eu"/>
+            <color key="backgroundColor" systemColor="systemGray6Color"/>
+            <constraints>
+                <constraint firstAttribute="trailing" secondItem="jr7-cp-sPd" secondAttribute="trailing" priority="750" constant="5" id="2rL-qx-Fbx"/>
+                <constraint firstItem="jr7-cp-sPd" firstAttribute="leading" secondItem="gTV-IL-0wX" secondAttribute="leading" constant="5" id="9hb-hL-HhS"/>
+                <constraint firstAttribute="bottom" secondItem="QfU-Gs-seW" secondAttribute="bottom" priority="750" constant="6" id="TWe-eR-PQ0"/>
+                <constraint firstItem="QfU-Gs-seW" firstAttribute="leading" secondItem="gTV-IL-0wX" secondAttribute="leading" constant="5" id="XVD-lE-0eF"/>
+                <constraint firstItem="QfU-Gs-seW" firstAttribute="top" secondItem="jr7-cp-sPd" secondAttribute="bottom" constant="5" id="YXU-FW-0ba"/>
+                <constraint firstAttribute="trailing" secondItem="QfU-Gs-seW" secondAttribute="trailing" priority="750" constant="5" id="aAi-h3-VCk"/>
+                <constraint firstItem="jr7-cp-sPd" firstAttribute="top" secondItem="gTV-IL-0wX" secondAttribute="top" constant="5" id="kXD-dA-9Jk"/>
+            </constraints>
+            <size key="customSize" width="103" height="101"/>
+            <connections>
+                <outlet property="recentImageView" destination="jr7-cp-sPd" id="ceF-xY-h4F"/>
+                <outlet property="recentLabel" destination="QfU-Gs-seW" id="dXL-jq-AJR"/>
+            </connections>
+            <point key="canvasLocation" x="169.56521739130437" y="148.36956521739131"/>
+        </collectionViewCell>
+    </objects>
+    <resources>
+        <systemColor name="systemGray6Color">
+            <color red="0.94901960784313721" green="0.94901960784313721" blue="0.96862745098039216" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+        </systemColor>
+    </resources>
+</document>
