--- GitIt/View/XIB/HistoryHeaderView.xib
@@ -1,91 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="18122" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina5_5" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="18093"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="System colors in document resources" minToolsVersion="11.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="iN0-l3-epB" customClass="HistoryHeaderView" customModule="GitIt" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="496" height="175"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" distribution="equalSpacing" spacing="15" translatesAutoresizingMaskIntoConstraints="NO" id="XzN-mg-YnF">
-                    <rect key="frame" x="15" y="15" width="466" height="145"/>
-                    <subviews>
-                        <stackView opaque="NO" contentMode="scaleToFill" distribution="equalSpacing" spacing="283" translatesAutoresizingMaskIntoConstraints="NO" id="Lyl-4k-5cb">
-                            <rect key="frame" x="0.0" y="0.0" width="466" height="30"/>
-                            <subviews>
-                                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="History" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="7RN-nk-vy9">
-                                    <rect key="frame" x="0.0" y="0.0" width="54" height="30"/>
-                                    <constraints>
-                                        <constraint firstAttribute="height" constant="30" id="Pxw-Sk-fk6"/>
-                                    </constraints>
-                                    <fontDescription key="fontDescription" type="boldSystem" pointSize="15"/>
-                                    <nil key="textColor"/>
-                                    <nil key="highlightedColor"/>
-                                </label>
-                                <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="fNs-HW-4sy">
-                                    <rect key="frame" x="436" y="0.0" width="30" height="30"/>
-                                    <constraints>
-                                        <constraint firstAttribute="height" constant="30" id="11x-GF-sms"/>
-                                        <constraint firstAttribute="width" constant="30" id="8NA-1T-bOP"/>
-                                    </constraints>
-                                    <state key="normal" image="xmark.circle.fill" catalog="system"/>
-                                    <connections>
-                                        <action selector="clearHistory:" destination="iN0-l3-epB" eventType="touchUpInside" id="sEI-8f-Luu"/>
-                                    </connections>
-                                </button>
-                            </subviews>
-                        </stackView>
-                        <collectionView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" alwaysBounceHorizontal="YES" pagingEnabled="YES" showsHorizontalScrollIndicator="NO" showsVerticalScrollIndicator="NO" dataMode="none" prefetchingEnabled="NO" translatesAutoresizingMaskIntoConstraints="NO" id="lvq-Mg-PP7">
-                            <rect key="frame" x="0.0" y="45" width="466" height="100"/>
-                            <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-                            <constraints>
-                                <constraint firstAttribute="height" constant="100" id="gIB-gs-3et"/>
-                            </constraints>
-                            <collectionViewFlowLayout key="collectionViewLayout" scrollDirection="horizontal" automaticEstimatedItemSize="YES" minimumLineSpacing="15" minimumInteritemSpacing="15" id="0Y0-TV-3b4">
-                                <size key="itemSize" width="75" height="100"/>
-                                <size key="headerReferenceSize" width="0.0" height="0.0"/>
-                                <size key="footerReferenceSize" width="0.0" height="0.0"/>
-                                <inset key="sectionInset" minX="0.0" minY="0.0" maxX="0.0" maxY="0.0"/>
-                            </collectionViewFlowLayout>
-                        </collectionView>
-                    </subviews>
-                    <constraints>
-                        <constraint firstAttribute="trailing" secondItem="Lyl-4k-5cb" secondAttribute="trailing" id="AHs-rN-0y3"/>
-                        <constraint firstItem="Lyl-4k-5cb" firstAttribute="top" secondItem="XzN-mg-YnF" secondAttribute="top" id="Nlc-nt-RbB"/>
-                        <constraint firstAttribute="trailing" secondItem="lvq-Mg-PP7" secondAttribute="trailing" id="c7u-zF-RHV"/>
-                        <constraint firstAttribute="bottom" secondItem="lvq-Mg-PP7" secondAttribute="bottom" id="esu-8e-tlL"/>
-                        <constraint firstItem="lvq-Mg-PP7" firstAttribute="leading" secondItem="XzN-mg-YnF" secondAttribute="leading" id="hgW-Ll-yiM"/>
-                        <constraint firstItem="Lyl-4k-5cb" firstAttribute="leading" secondItem="XzN-mg-YnF" secondAttribute="leading" id="snt-KZ-69x"/>
-                    </constraints>
-                </stackView>
-            </subviews>
-            <viewLayoutGuide key="safeArea" id="vUN-kp-3ea"/>
-            <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-            <constraints>
-                <constraint firstItem="XzN-mg-YnF" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="15" id="4UB-c3-2Ct"/>
-                <constraint firstItem="vUN-kp-3ea" firstAttribute="bottom" secondItem="XzN-mg-YnF" secondAttribute="bottom" priority="750" constant="15" id="Dno-Jb-SCO"/>
-                <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" secondItem="XzN-mg-YnF" secondAttribute="trailing" priority="750" constant="15" id="kBm-Cm-Q5a"/>
-                <constraint firstItem="XzN-mg-YnF" firstAttribute="top" secondItem="vUN-kp-3ea" secondAttribute="top" constant="15" id="uaH-Qx-UlK"/>
-            </constraints>
-            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
-            <connections>
-                <outlet property="collectionView" destination="lvq-Mg-PP7" id="UnL-a2-9or"/>
-            </connections>
-            <point key="canvasLocation" x="155.07246376811597" y="-103.12500000000001"/>
-        </view>
-    </objects>
-    <resources>
-        <image name="xmark.circle.fill" catalog="system" width="128" height="121"/>
-        <systemColor name="systemBackgroundColor">
-            <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-        </systemColor>
-    </resources>
-</document>
