--- GitIt/View/XIB/HistoryTableViewCell.xib
@@ -1,61 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="18122" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina6_1" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="18093"/>
-        <capability name="Named colors" minToolsVersion="9.0"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" reuseIdentifier="HistoryTableViewCell" rowHeight="50" id="KGk-i7-Jjw" customClass="HistoryTableViewCell" customModule="GitIt" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="414" height="50"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
-                <rect key="frame" x="0.0" y="0.0" width="414" height="50"/>
-                <autoresizingMask key="autoresizingMask"/>
-                <subviews>
-                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" verticalHuggingPriority="251" text="Recent" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="Sz1-oA-juU">
-                        <rect key="frame" x="20" y="15" width="56.5" height="20"/>
-                        <constraints>
-                            <constraint firstAttribute="height" constant="20" id="nzj-uj-t5d"/>
-                        </constraints>
-                        <fontDescription key="fontDescription" type="system" pointSize="18"/>
-                        <nil key="textColor"/>
-                        <nil key="highlightedColor"/>
-                    </label>
-                    <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="arrow.up.left" catalog="system" translatesAutoresizingMaskIntoConstraints="NO" id="hI1-6U-WHg">
-                        <rect key="frame" x="379" y="17.5" width="20" height="14.5"/>
-                        <color key="tintColor" name="AccentColor"/>
-                        <constraints>
-                            <constraint firstAttribute="height" constant="20" id="K9X-Y9-ZBS"/>
-                            <constraint firstAttribute="width" constant="20" id="nA7-cY-eMu"/>
-                        </constraints>
-                    </imageView>
-                </subviews>
-                <constraints>
-                    <constraint firstAttribute="trailing" secondItem="hI1-6U-WHg" secondAttribute="trailing" constant="15" id="HdJ-b7-0ED"/>
-                    <constraint firstItem="Sz1-oA-juU" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="15" id="OUH-Pq-A2Y"/>
-                    <constraint firstAttribute="bottom" secondItem="Sz1-oA-juU" secondAttribute="bottom" priority="750" constant="15" id="UYp-qB-KNP"/>
-                    <constraint firstAttribute="bottom" secondItem="hI1-6U-WHg" secondAttribute="bottom" priority="750" constant="15" id="hOZ-8l-rsv"/>
-                    <constraint firstItem="hI1-6U-WHg" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="15" id="kxG-ao-mVF"/>
-                    <constraint firstItem="Sz1-oA-juU" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" constant="20" symbolic="YES" id="uw1-1d-Zp2"/>
-                </constraints>
-            </tableViewCellContentView>
-            <viewLayoutGuide key="safeArea" id="njF-e1-oar"/>
-            <connections>
-                <outlet property="recentsLabel" destination="Sz1-oA-juU" id="S4D-9z-TfY"/>
-            </connections>
-            <point key="canvasLocation" x="147.82608695652175" y="140.95982142857142"/>
-        </tableViewCell>
-    </objects>
-    <resources>
-        <image name="arrow.up.left" catalog="system" width="128" height="112"/>
-        <namedColor name="AccentColor">
-            <color red="0.0" green="0.46000000000000002" blue="0.89000000000000001" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-        </namedColor>
-    </resources>
-</document>
