--- GitIt/View/XIB/UserTableViewCell.xib
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="18122" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina5_5" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="18093"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell contentMode="scaleToFill" selectionStyle="default" accessoryType="disclosureIndicator" indentationWidth="10" reuseIdentifier="UserTableViewCell" id="KGk-i7-Jjw" customClass="UserTableViewCell" customModule="GitIt" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="414" height="74"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
-                <rect key="frame" x="0.0" y="0.0" width="384.33333333333331" height="74"/>
-                <autoresizingMask key="autoresizingMask"/>
-                <subviews>
-                    <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="PDQ-al-liA" customClass="AsynchronousUIImageView" customModule="GitIt" customModuleProvider="target">
-                        <rect key="frame" x="10" y="5" width="64" height="64"/>
-                        <constraints>
-                            <constraint firstAttribute="width" constant="64" id="Fb7-bZ-CRU"/>
-                            <constraint firstAttribute="height" constant="64" id="ZLc-sq-R3f"/>
-                        </constraints>
-                    </imageView>
-                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="User Login" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="LLK-Ad-4I7">
-                        <rect key="frame" x="84.000000000000014" y="12" width="192.33333333333337" height="50"/>
-                        <constraints>
-                            <constraint firstAttribute="height" constant="50" id="eOA-m0-dg9"/>
-                        </constraints>
-                        <fontDescription key="fontDescription" type="boldSystem" pointSize="20"/>
-                        <nil key="textColor"/>
-                        <nil key="highlightedColor"/>
-                    </label>
-                </subviews>
-                <constraints>
-                    <constraint firstItem="PDQ-al-liA" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="5" id="6yS-nb-c1Q"/>
-                    <constraint firstAttribute="bottom" secondItem="PDQ-al-liA" secondAttribute="bottom" priority="750" constant="5" id="9Co-ss-Wxi"/>
-                    <constraint firstItem="PDQ-al-liA" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" constant="10" id="JBy-rQ-UFp"/>
-                    <constraint firstItem="LLK-Ad-4I7" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="topMargin" constant="1" id="gj9-aE-4bg"/>
-                    <constraint firstAttribute="bottomMargin" secondItem="LLK-Ad-4I7" secondAttribute="bottom" priority="750" constant="1" id="jkf-nY-486"/>
-                    <constraint firstAttribute="trailingMargin" secondItem="LLK-Ad-4I7" secondAttribute="trailing" constant="100" id="sO6-o7-1Ml"/>
-                    <constraint firstItem="LLK-Ad-4I7" firstAttribute="leading" secondItem="PDQ-al-liA" secondAttribute="trailing" constant="10" id="z3g-0n-MSv"/>
-                </constraints>
-            </tableViewCellContentView>
-            <viewLayoutGuide key="safeArea" id="njF-e1-oar"/>
-            <inset key="separatorInset" minX="64" minY="0.0" maxX="0.0" maxY="0.0"/>
-            <connections>
-                <outlet property="userImageView" destination="PDQ-al-liA" id="8MF-4r-Xrz"/>
-                <outlet property="userLoginLabel" destination="LLK-Ad-4I7" id="GWe-aM-3ho"/>
-            </connections>
-            <point key="canvasLocation" x="134.78260869565219" y="141.03260869565219"/>
-        </tableViewCell>
-    </objects>
-</document>
